Added the logo in various widths and formats
The logo has been generated into files so that each one contains a combination of the stylized letter 'B' in one of the three brand colors atop a complimentary background also created from the same set of three colors. The logo files have been placed into folders that denote their width if they are in .png format, or an .svg folder specifically for that undefined size layout.